requirements.txt
absl-py==0.10.0
astor==0.8.1
attrs==20.1.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
clickclick==1.2.2
connexion==2.6.0
cycler==0.10.0
Flask==1.1.2
Flask-Cors==3.0.8
flask-marshmallow==0.11.0
Flask-SQLAlchemy==2.4.1
gast==0.2.2
google-pasta==0.2.0
grpcio==1.31.0
h5py==2.10.0
idna==2.8
importlib-metadata==1.7.0
inflection==0.5.1
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.14.1
jsonschema==3.2.0
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
Markdown==3.2.2
MarkupSafe==1.1.1
marshmallow==3.7.1
marshmallow-sqlalchemy==0.22.3
matplotlib==3.1.1
numpy==1.19.1
openapi-spec-validator==0.2.9
opt-einsum==3.3.0
pandas==0.25.3
Pillow==7.1.2
protobuf==3.13.0
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.22.0
scipy==1.5.2
six==1.15.0
scikit-learn==0.22.1 
SQLAlchemy==1.3.19
swagger-ui-bundle==0.0.6
tensorboard==1.15.0
tensorflow==1.15.0
tensorflow-estimator==1.15.1
tensorflow-serving-api==1.11.0
termcolor==1.1.0
urllib3==1.25.10
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
